feat(Matterport): Integrate Matterport Showcase Viewer into TwinMaker SceneComposer
**EXPERIMENTAL**: This change is still in active development, and APIs are not yet finalized. Use at your own risk.

This change includes a brand new capability for the iot-app-kit scene-composer component,
which allows customers to add a layer to their digital twin that renders a scene from Matterport.

It also enables the customer to view the combined TwinMaker + Matterport scene in the Matterport
Showcase viewer which enables some pretty cool features:

- Navigation is simpler and more intuitive
- Immersive View support, browser your scene from the first person perspective!

There are more changes to come, so for now this is behind an experimental flag. Use at your own risk,
as we will likely make breaking changes to this feature over the next few weeks.

Gated behind COMPOSER_FEATURES.Matterport, and requires a Matterport ModelId and applicationKey to work.